# swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "App Store Connect Adapters",
    "description": "HTTP API for querying an App Store Connect adapters host.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:48189"
    }
  ],
  "paths": {
...
    "/api/app-store-connect/v1.0/asset-model/{adapterId}/browse": {
      "get": {
        "tags": [
          "AssetModelBrowser"
        ],
        "summary": "Browses the asset model hierarchy. Up to MaxNodesPerQuery will be returned.",
        "operationId": "AssetModelBrowser_BrowseNodes",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID to query.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "start",
            "in": "query",
            "description": "The optional starting node ID. If no value is specified, browsing will start at the top-level nodes in the hierarchy.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "page",
            "in": "query",
            "description": "The results page to retrieve.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            },
            "x-position": 3
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The page size for the query.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "Successful responses contain the matching AssetModelNode objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetModelNode"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AssetModelBrowser"
        ],
        "summary": "Browses the asset model hierarchy. Up to MaxNodesPerQuery will be returned.",
        "operationId": "AssetModelBrowser_BrowseNodesPost",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID to query.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The browse request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowseAssetModelNodesRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "Successful responses contain the matching AssetModelNode objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetModelNode"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/app-store-connect/v1.0/asset-model/{adapterId}/get-by-id": {
      "post": {
        "tags": [
          "AssetModelBrowser"
        ],
        "summary": "Gets a collection of nodes by ID. Up to MaxNodesPerQuery will be returned.",
        "operationId": "AssetModelBrowser_GetNodes",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID to query.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The request object describing the nodes to retrieve.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAssetModelNodesRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "Successful responses contain the matching AssetModelNode objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetModelNode"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/app-store-connect/v1.0/asset-model/{adapterId}/find": {
      "post": {
        "tags": [
          "AssetModelBrowser"
        ],
        "summary": "Finds nodes matching the specified search filter. Up to MaxNodesPerQuery will be returned.",
        "operationId": "AssetModelBrowser_FindNodes",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID to query.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The request object describing the nodes to retrieve.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindAssetModelNodesRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "Successful responses contain the matching AssetModelNode objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetModelNode"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PageableAdapterRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AdapterRequest"
          },
          {
            "type": "object",
            "description": "Base class for adapter requests that support paging.",
            "x-abstract": true,
            "additionalProperties": false,
            "properties": {
              "pageSize": {
                "type": "integer",
                "description": "The page size for the query.",
                "format": "int32",
                "default": 10,
                "maximum": 2147483647.0,
                "minimum": 1.0
              },
              "page": {
                "type": "integer",
                "description": "The page number for the query.",
                "format": "int32",
                "default": 1,
                "maximum": 2147483647.0,
                "minimum": 1.0
              }
            }
          }
        ]
      },
      "AdapterRequest": {
        "type": "object",
        "description": "Base class that adapter request objects can inherit from.",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "properties": {
            "type": "object",
            "description": "Additional request properties. These can be used to provide bespoke query parameters \nsupported by the adapter.",
            "nullable": true,
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "AdapterProperty": {
        "type": "object",
        "description": "Describes a custom property associated with an adapter, tag, tag value, event message, etc.",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The property name.",
            "minLength": 1
          },
          "value": {
            "description": "The property value.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Variant"
              }
            ]
          },
          "description": {
            "type": "string",
            "description": "The property description.",
            "nullable": true
          }
        }
      },
      "Variant": {
        "type": "object",
        "description": "Describes a variant value.",
        "additionalProperties": false,
        "properties": {
          "value": {
            "description": "The value.",
            "nullable": true
          },
          "type": {
            "description": "The variant type.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/VariantType"
              }
            ]
          }
        }
      },
      "VariantType": {
        "type": "string",
        "description": "Describes the type of a variant value.",
        "x-enumNames": [
          "Unknown",
          "Null",
          "Object",
          "Boolean",
          "SByte",
          "Byte",
          "Int16",
          "UInt16",
          "Int32",
          "UInt32",
          "Int64",
          "UInt64",
          "Float",
          "Double",
          "String",
          "DateTime",
          "TimeSpan",
          "Url"
        ],
        "enum": [
          "Unknown",
          "Null",
          "Object",
          "Boolean",
          "SByte",
          "Byte",
          "Int16",
          "UInt16",
          "Int32",
          "UInt32",
          "Int64",
          "UInt64",
          "Float",
          "Double",
          "String",
          "DateTime",
          "TimeSpan",
          "Url"
        ]
      },
...
      "AssetModelNode": {
        "type": "object",
        "description": "Describes a node in an adapter's asset model.",
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier for the node.",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The node name.",
            "minLength": 1
          },
          "nodeType": {
            "description": "The type of the node.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/NodeType"
              }
            ]
          },
          "nodeSubType": {
            "type": "string",
            "description": "The subtype of the node. This is a text value supplied by the adapter, intended to \nprovide more information about the node type (e.g. if the node represents a conceptual \nitem such as a folder in a hierarchical structure).",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "The node description.",
            "nullable": true
          },
          "parent": {
            "type": "string",
            "description": "The parent ID of the node. Top-level nodes will have a value of null.",
            "nullable": true
          },
          "hasChildren": {
            "type": "boolean",
            "description": "Indicates if this node has any child nodes."
          },
          "dataReference": {
            "description": "The data reference associated with the node.",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/DataReference"
              }
            ]
          },
          "properties": {
            "type": "array",
            "description": "Additional properties associated with the node.",
            "items": {
              "$ref": "#/components/schemas/AdapterProperty"
            }
          }
        }
      },
      "NodeType": {
        "type": "string",
        "description": "Describes the type of a node in an asset model.",
        "x-enumNames": [
          "Unknown",
          "Object",
          "Variable",
          "ObjectType",
          "VariableType",
          "Other"
        ],
        "enum": [
          "Unknown",
          "Object",
          "Variable",
          "ObjectType",
          "VariableType",
          "Other"
        ]
      },
      "DataReference": {
        "type": "object",
        "description": "Describes a data reference on an AssetModelNode. Note that the reference can \nbe to a tag on a different adapter.",
        "additionalProperties": false,
        "properties": {
          "adapterId": {
            "type": "string",
            "description": "The adapter ID for the reference."
          },
          "tag": {
            "description": "The tag identifier for the reference.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/TagIdentifier"
              }
            ]
          }
        }
      },
      "TagIdentifier": {
        "type": "object",
        "description": "Defines basic information for identifying a real-time data tag.",
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier for the tag.",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The tag name.",
            "minLength": 1
          }
        }
      },
      "BrowseAssetModelNodesRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PageableAdapterRequest"
          },
          {
            "type": "object",
            "description": "Describes a request to browse nodes in an adapter's asset model.",
            "additionalProperties": false,
            "properties": {
              "parentId": {
                "type": "string",
                "description": "The ID of the parent node to start at. Specify null to request top-level \nnodes.",
                "nullable": true
              }
            }
          }
        ]
      },
      "GetAssetModelNodesRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AdapterRequest"
          },
          {
            "type": "object",
            "description": "Describes a request to get asset model nodes by ID.",
            "additionalProperties": false,
            "required": [
              "nodes"
            ],
            "properties": {
              "nodes": {
                "type": "array",
                "description": "The IDs of the nodes to retrieve.",
                "minItems": 1,
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "FindAssetModelNodesRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PageableAdapterRequest"
          },
          {
            "type": "object",
            "description": "Describes a request to search for asset model nodes.",
            "additionalProperties": false,
            "properties": {
              "name": {
                "type": "string",
                "description": "The name filter.",
                "nullable": true
              },
              "description": {
                "type": "string",
                "description": "The description filter.",
                "nullable": true
              }
            }
          }
        ]
      },
...
    }
  }
}